+++ src/lib/components/events/UbicationTicketsCard.svelte
@@ -66,19 +66,26 @@
 </div> -->
 
 <div class:opacity-30={ticket.amount === 0} class="w-full">
-  <div class="relative bg-primary p-6 text-black" >
+  <div class="relative bg-primary p-6 text-black h-full" >
+
+    <div 
+        class="absolute inset-0 opacity-90"
+        style="background-image: url('https://hebbkx1anhila5yf.public.blob.vercel-storage.com/crumpled-paper-KLaPwXItbh5Sj09tzbCcSbD7qsWJBN.webp'); background-size: cover; background-position: center; mix-blend-mode: multiply;"
+      ></div>
+
     <!-- Notches -->
     <div class="absolute left-0 top-16 w-4 h-4 bg-black" />
     <div class="absolute right-0 top-16 w-4 h-4 bg-black" />
 
-    <h2 class="text-2xl font-bold mb-2 text-center pb-2">{ticket.name}</h2>
+    <h2 class="text-3xl font-ibm italic mb-2 text-center pb-2">Ticket {ticket.name}</h2>
     <div class="w-full border border-black/30 border-b-0 border-dashed"></div>
-    <div class="flex justify-center gap-4 items-center mb-4 mt-2">
+    <div class="flex justify-center gap-4 items-center mt-4">
       <p class="text-2xl font-bold text-center line-through opacity-50">
         {dicountPercentage ? `$${priceBeforeDiscount}` : ""}
       </p>
-      <p class="text-2xl font-bold text-center">{formatedPrice || 0}</p>
+      <p class="text-3xl text-center">{formatedPrice || 0}</p>
     </div>
+    {#if ticket.amount <= 15}
     <p
       class={`mt-2 font-semibold tracking-widest uppercase text-sm text-center`}
     >
@@ -86,5 +93,6 @@
         ? `Últimas ${ticket.amount} entradas`
         : "Agotado"}
     </p>
+    {/if}
   </div>
 </div>